refactor: remove migrated feature flag usage
