+++ operate/client/src/modules/hooks/flowNodeMetadata.ts
@@ -6,51 +6,14 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {flowNodeSelectionStore} from 'modules/stores/flowNodeSelection';
 import {
   useHasRunningOrFinishedTokens,
   useNewTokenCountForSelectedNode,
 } from './flowNodeSelection';
 import {useElementInstancesCount} from './useElementInstancesCount';
 import {useProcessInstanceElementSelection} from './useProcessInstanceElementSelection';
-import {IS_ELEMENT_SELECTION_V2} from 'modules/feature-flags';
 
 const useHasMultipleInstances = () => {
-  const hasMultipleInstancesV2 = useHasMultipleInstancesV2();
-  const hasMultipleInstancesV1 = useHasMultipleInstancesV1();
-
-  if (IS_ELEMENT_SELECTION_V2) {
-    return hasMultipleInstancesV2;
-  }
-
-  return hasMultipleInstancesV1;
-};
-
-const useHasMultipleInstancesV1 = () => {
-  const hasRunningOrFinishedTokens = useHasRunningOrFinishedTokens();
-  const newTokenCountForSelectedNode = useNewTokenCountForSelectedNode();
-  const elementInstancesCount = useElementInstancesCount(
-    flowNodeSelectionStore.state.selection?.flowNodeId,
-  );
-
-  if (flowNodeSelectionStore.state.selection?.isMultiInstance) {
-    return false;
-  }
-
-  if (
-    flowNodeSelectionStore.state.selection?.flowNodeInstanceId !== undefined
-  ) {
-    return false;
-  }
-
-  if (!hasRunningOrFinishedTokens) {
-    return newTokenCountForSelectedNode > 1;
-  }
-
-  return (elementInstancesCount ?? 0) > 1 || newTokenCountForSelectedNode > 0;
-};
-
-const useHasMultipleInstancesV2 = () => {
   const hasRunningOrFinishedTokens = useHasRunningOrFinishedTokens();
   const newTokenCountForSelectedNode = useNewTokenCountForSelectedNode();
   const {
