+++ operate/client/src/modules/hooks/variables.ts
@@ -18,8 +18,6 @@ import {useHasMultipleInstances} from './flowNodeMetadata';
 import {useProcessInstanceElementSelection} from './useProcessInstanceElementSelection';
 import {TOKEN_OPERATIONS} from 'modules/constants';
 import {useElementSelectionInstanceKey} from './useElementSelectionInstanceKey';
-import {IS_ELEMENT_SELECTION_V2} from 'modules/feature-flags';
-import {flowNodeSelectionStore} from 'modules/stores/flowNodeSelection';
 
 const useHasNoContent = () => {
   const newTokenCountForSelectedNode = useNewTokenCountForSelectedNode();
@@ -76,18 +74,8 @@ const useDisplayStatus = ({
 }) => {
   const hasNoContent = useHasNoContent();
   const hasMultipleInstances = useHasMultipleInstances();
-  const newTokenCountForSelectedNode = useNewTokenCountForSelectedNode();
-  let isPlaceholderSelected = useIsPlaceholderSelected();
-  let {isFetchingElementError} = useProcessInstanceElementSelection();
-
-  // TODO: Remove these assignments once the feature flag is fully rolled out
-  isFetchingElementError = IS_ELEMENT_SELECTION_V2
-    ? isFetchingElementError
-    : flowNodeMetaDataStore.state.status === 'error';
-  isPlaceholderSelected = IS_ELEMENT_SELECTION_V2
-    ? isPlaceholderSelected
-    : flowNodeSelectionStore.state.selection?.isPlaceholder ||
-      newTokenCountForSelectedNode === 1;
+  const isPlaceholderSelected = useIsPlaceholderSelected();
+  const {isFetchingElementError} = useProcessInstanceElementSelection();
 
   if (isError || isFetchingElementError) {
     return 'error';
